Removing logs for files scans. changing payload collection method for effective traversal. Fix for Stuck stop button by ensuring quick job cancellation. Removed redundant tests
